@@ -18,33 +18,33 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^20.0.0",
-		"@angular/common": "^20.0.0",
-		"@angular/compiler": "^20.0.0",
-		"@angular/core": "^20.0.0",
-		"@angular/forms": "^20.0.0",
-		"@angular/platform-browser": "^20.0.0",
-		"@angular/platform-browser-dynamic": "^20.0.0",
-		"@angular/platform-server": "^20.0.0",
-		"@angular/router": "^20.0.0",
-		"@angular/ssr": "^20.0.0",
+		"@angular/animations": "^21.0.0",
+		"@angular/common": "^21.0.0",
+		"@angular/compiler": "^21.0.0",
+		"@angular/core": "^21.0.0",
+		"@angular/forms": "^21.0.0",
+		"@angular/platform-browser": "^21.0.0",
+		"@angular/platform-browser-dynamic": "^21.0.0",
+		"@angular/platform-server": "^21.0.0",
+		"@angular/router": "^21.0.0",
+		"@angular/ssr": "^21.0.0",
 		"@ngrx/signals": "^20.0.0",
 		"ngx-toastr": "^19.0.0",
 		"rxjs": "~7.8.2",
-		"tslib": "^2.3.0",
-		"zone.js": "~0.15.1"
+		"tslib": "^2.3.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "^20.0.0",
-		"@angular/cli": "^20.0.0",
-		"@angular/compiler-cli": "^20.0.0",
+		"@angular/build": "^21.0.0",
+		"@angular/cli": "^21.0.0",
+		"@angular/compiler-cli": "^21.0.0",
 		"@biomejs/biome": "2.3.7",
 		"@cloudflare/workers-types": "^4.20241004.0",
 		"@cypress/schematic": "4.1.1",
 		"@miniflare/tre": "^3.0.0-next.13",
 		"@types/jasmine": "~5.1.0",
-		"@types/node": "^18.18.0",
+		"@types/node": "^24.10.1",
 		"cypress": "latest",
+		"esbuild": "^0.27.0",
 		"jasmine-core": "~5.2.0",
 		"karma": "~6.4.0",
 		"karma-chrome-launcher": "~3.2.0",
@@ -52,7 +52,7 @@
 		"karma-jasmine": "~5.1.0",
 		"karma-jasmine-html-reporter": "~2.1.0",
 		"tailwindcss": "^3.4.13",
-		"typescript": "~5.8.0",
-		"wrangler": "^3.107.2"
+		"typescript": "~5.9.3",
+		"wrangler": "^4.50.0"
 	}
 }
