@@ -1,7 +1,19 @@
-import { bootstrapApplication } from '@angular/platform-browser';
+import { provideZonelessChangeDetection } from '@angular/core';
+import {
+	BootstrapContext,
+	bootstrapApplication,
+} from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { config } from './app/app.config.server';
 
-const bootstrap = () => bootstrapApplication(AppComponent, config);
+const bootstrap = (context: BootstrapContext) =>
+	bootstrapApplication(
+		AppComponent,
+		{
+			...config,
+			providers: [provideZonelessChangeDetection(), ...config.providers],
+		},
+		context,
+	);
 
 export default bootstrap;
